Improve kmapcontent
