+++ KMaPDemo/src/commonMain/kotlin/com/rafambn/kmap/App.kt
@@ -69,87 +69,3 @@ fun App() = AppTheme {
         }
     }
 }
-
-//@Composable
-// fun App() = AppTheme {
-//    Surface(modifier = Modifier.fillMaxSize()) {
-//        Box {
-//            val motionController = rememberMotionController()
-//            val mapState = rememberMapState(mapProperties = OSMMapProperties())
-//            KMaP(
-//                modifier = Modifier.align(Alignment.Center).size(300.dp, 600.dp),
-//                motionController = motionController,
-//                mapState = mapState,
-//                canvasGestureListener = DefaultCanvasGestureListener()
-//            ) {
-//                canvas(tileSource = OSMTileSource("com.rafambn.kmap")::getTile)
-//                markers(
-//                    listOf(
-//                        MarkerParameters(
-//                            ProjectedCoordinates(-45.949303, -21.424608),
-//                            drawPosition = DrawPosition.BOTTOM_RIGHT,
-//                            rotateWithMap = true,
-//                            tag = "zika"
-//                        ),
-//                        MarkerParameters(
-//                            ProjectedCoordinates(-46.949303, -21.424608),
-//                            drawPosition = DrawPosition.BOTTOM_RIGHT,
-//                            rotateWithMap = true,
-//                            tag = "zika"
-//                        ),
-//                        MarkerParameters(
-//                            ProjectedCoordinates(180.0, -85.0),
-//                            drawPosition = DrawPosition.BOTTOM_RIGHT,
-//                            rotateWithMap = true,
-//                            tag = "zika"
-//                        ),
-//                    )
-//                ) {
-//                    Image(
-//                        painterResource(Res.drawable.teste),
-//                        "fd",
-//                        Modifier
-//                            .background(Color.Black)
-//                            .size(32.dp)
-//                            .clickable {
-//                                println("fsdfd")
-//                            }
-//                    )
-//                }
-//                markers(
-//                    listOf(
-//                        MarkerParameters(
-//                            ProjectedCoordinates(180.0, 85.0),
-//                            drawPosition = DrawPosition.TOP_RIGHT,
-//                            rotateWithMap = false,
-//                        )
-//                    )
-//                ) {
-//                    val cor = remember { mutableStateOf(Color.Black) }
-//                    val rnd = remember { Random(545) }
-//                    Box(
-//                        Modifier
-//                            .background(cor.value)
-//                            .size(32.dp)
-//                            .clickable {
-//                                cor.value = Color(rnd.nextInt(256), rnd.nextInt(256), rnd.nextInt(256), rnd.nextInt(256))
-//                            }
-//                    ) {
-//                    }
-//                }
-//                cluster(ClusterParameters("zika", 50.dp, rotateWithMap = true)) {
-//                    Image(
-//                        painterResource(Res.drawable.teste2),
-//                        "fd",
-//                        Modifier
-//                            .background(Color.Black)
-//                            .size(32.dp)
-//                            .clickable {
-//                                println("fsdfd")
-//                            }
-//                    )
-//                }
-//            }
-//        }
-//    }
-//}
