+++ KMaPDemo/src/commonMain/kotlin/com/rafambn/kmap/screens/LayersMap.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.rafambn.kmap.DefaultCanvasGestureListener
 import com.rafambn.kmap.KMaP
+import com.rafambn.kmap.canvas
 import com.rafambn.kmap.config.characteristics.TileSource
 import com.rafambn.kmap.core.CanvasParameters
 import com.rafambn.kmap.core.rememberMotionController
