+++ KMaPDemo/src/commonMain/kotlin/com/rafambn/kmap/screens/Markers.kt
@@ -19,16 +19,17 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.rafambn.kmap.DefaultCanvasGestureListener
 import com.rafambn.kmap.KMaP
+import com.rafambn.kmap.canvas
 import com.rafambn.kmap.core.DrawPosition
 import com.rafambn.kmap.core.MarkerParameters
 import com.rafambn.kmap.core.rememberMotionController
 import com.rafambn.kmap.core.state.rememberMapState
+import com.rafambn.kmap.marker
+import com.rafambn.kmap.markers
 import com.rafambn.kmap.utils.offsets.ProjectedCoordinates
 import kmap.kmapdemo.generated.resources.Res
 import kmap.kmapdemo.generated.resources.back_arrow
 import kmap.kmapdemo.generated.resources.pin
-import kmap.kmapdemo.generated.resources.teste
-import kmap.kmapdemo.generated.resources.teste2
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.vectorResource
 import kotlin.random.Random
@@ -41,6 +42,7 @@ fun MarkerMapRoot(
         val motionController = rememberMotionController()
         val mapState = rememberMapState(mapProperties = SimpleMapProperties())
         val markersList = remember { mutableStateListOf<MarkerParameters>() }
+        println("newState ${markersList.size}")
         KMaP(
             modifier = Modifier.align(Alignment.Center).fillMaxSize(),
             motionController = motionController,
@@ -48,68 +50,63 @@ fun MarkerMapRoot(
             canvasGestureListener = DefaultCanvasGestureListener()
         ) {
             canvas(tileSource = SimpleMapTileSource()::getTile)
-            markers(
-                listOf(
-                    MarkerParameters(
-                        ProjectedCoordinates(-0.0, -0.0),
-                        drawPosition = DrawPosition.TOP_RIGHT,
-                        scaleWithMap = false,
-                        tag = "Fixed size"
-                    )
+            marker(
+                MarkerParameters(
+                    ProjectedCoordinates(-0.0, -0.0),
+                    drawPosition = DrawPosition.TOP_RIGHT,
+                    scaleWithMap = false,
+                    tag = "Fixed size"
                 )
             ) {
                 Text(
                     text = "Fixed size",
                     modifier = Modifier
                         .background(Color.Black)
-                        .padding(16.dp)
+                        .padding(16.dp),
+                    color = Color.White
                 )
             }
-            markers(
-                listOf(
-                    MarkerParameters(
-                        ProjectedCoordinates(-0.0, -10.0),
-                        drawPosition = DrawPosition.TOP_RIGHT,
-                        scaleWithMap = true,
-                        zoomToFix = 1F,
-                        tag = "Scale with zoom"
-                    )
+            marker(
+                MarkerParameters(
+                    ProjectedCoordinates(-0.0, -10.0),
+                    drawPosition = DrawPosition.TOP_RIGHT,
+                    scaleWithMap = true,
+                    zoomToFix = 1F,
+                    tag = "Scale with zoom"
                 )
             ) {
                 Text(
                     text = "Scale with zoom",
                     modifier = Modifier
                         .background(Color.Black)
-                        .padding(16.dp)
+                        .padding(16.dp),
+                    color = Color.White
                 )
             }
-            markers(
-                listOf(
-                    MarkerParameters(
-                        ProjectedCoordinates(-0.0, 20.0),
-                        drawPosition = DrawPosition.TOP_RIGHT,
-                        scaleWithMap = false,
-                        rotateWithMap = true,
-                        rotation = -45.0,
-                        tag = "Rotate with map and custom start angle"
-                    )
+            marker(
+                MarkerParameters(
+                    ProjectedCoordinates(-0.0, 20.0),
+                    drawPosition = DrawPosition.TOP_RIGHT,
+                    scaleWithMap = false,
+                    rotateWithMap = true,
+                    rotation = -45.0,
+                    tag = "Rotate with map and custom start angle"
                 )
             ) {
                 Text(
                     text = "Rotate with map",
                     modifier = Modifier
                         .background(Color.Black)
-                        .padding(16.dp)
+                        .padding(16.dp),
+                    color = Color.White
                 )
             }
-            markers(
-                listOf(
-                    MarkerParameters(
-                        ProjectedCoordinates(-90.0, 0.0),
-                        drawPosition = DrawPosition.TOP_RIGHT,
-                        scaleWithMap = false,
-                        tag = "Clickable"
-                    )
+            marker(
+                MarkerParameters(
+                    ProjectedCoordinates(-90.0, 0.0),
+                    drawPosition = DrawPosition.TOP_RIGHT,
+                    scaleWithMap = false,
+                    tag = "Clickable"
                 )
             ) {
                 val cor = remember { mutableStateOf(Color.Black) }
@@ -128,10 +125,11 @@ fun MarkerMapRoot(
                             count.value++
                         }
                         .background(cor.value)
-                        .padding(16.dp)
+                        .padding(16.dp),
+                    color = Color.Red
                 )
             }
-            markers(markersList) {//TODO fix/improve this
+            markers(markersList) {
                 Image(
                     painter = painterResource(Res.drawable.pin),
                     contentDescription = "Removable marker",
@@ -145,13 +143,13 @@ fun MarkerMapRoot(
             onClick = {
                 markersList.add(
                     MarkerParameters(
-                        ProjectedCoordinates(90.0, 45.0),
+                        ProjectedCoordinates(90.0, 0.0),
                         drawPosition = DrawPosition.TOP_RIGHT,
                         scaleWithMap = false,
                         tag = "Removable marker"
                     )
                 )
-                println(markersList.size)
+                println("bottom clicked")
             },
             modifier = Modifier.align(Alignment.TopEnd),
         ) {
